5.3.0 服务器时间 vpn状态显示优化

+++ src/main/domain/sys/sys.systemd.service.ts
@@ -229,6 +229,9 @@ export class SysSystemdService {
 
 
     async delete_sys_systemd(name:string) {
+        if(name.includes(" ")) {
+            throw "error name";
+        }
         const sdtout = execSync(`systemctl show ${name} --property=FragmentPath`).toString();
         const filePath = sdtout.split("=")[1].replaceAll("\n","");
         execSync(`sudo systemctl stop ${name}`);
